+++ source/renderer/app/stores/NetworkStatusStore.ts
@@ -805,14 +805,10 @@ export default class NetworkStatusStore extends Store {
     return Promise.resolve();
   };
 
-  @action _onBlockSyncProgressUpdate = async ({
-    progress,
-    type,
-  }: GetBlockSyncProgressMainResponse) => {
-    this.blockSyncProgress = {
-      ...this.blockSyncProgress,
-      [type]: progress,
-    };
+  @action _onBlockSyncProgressUpdate = async (
+    blockSyncProgress: GetBlockSyncProgressMainResponse
+  ) => {
+    this.blockSyncProgress = blockSyncProgress;
   };
 
   @action